+++ examples/arduino/board/board_static_config/esp_panel_board_supported_conf.h
@@ -93,6 +93,7 @@
  * Waveshare Supported Boards (https://www.waveshare.com/):
  *
  *  -BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85 (ESP32_S3_Touch_LCD_1_85): https://www.waveshare.com/esp32-s3-touch-lcd-1.85.htm
+ *  -BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85_C (ESP32_S3_Touch_LCD_1_85_C): https://www.waveshare.com/esp32-s3-touch-lcd-1.85c.htm
  *  -BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_2_1 (ESP32_S3_Touch_LCD_2_1): https://www.waveshare.com/esp32-s3-touch-lcd-2.1.htm
  *  -BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3 (ESP32_S3_Touch_LCD_4_3): https://www.waveshare.com/esp32-s3-touch-lcd-4.3.htm
  *  -BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3_B (ESP32_S3_Touch_LCD_4_3_B): https://www.waveshare.com/esp32-s3-touch-lcd-4.3B.htm
@@ -102,6 +103,7 @@
  *  -BOARD_WAVESHARE_ESP32_P4_NANO (ESP32_P4_NANO): https://www.waveshare.com/esp32-p4-nano.htm
  */
 // #define BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85
+// #define BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85_C
 // #define BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_2_1
 // #define BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3
 // #define BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3_B